cephadm: move CephadmcOntext to cephadmlib/context.py
Pair-programmed-with: Adam King <adking@redhat.com>
